agents/requirements.txt
autogen==1.0.16
beartype==0.18.5
beautifulsoup4==4.12.3
common==0.1.2
duckduckgo_search==6.1.12
evaluate==0.4.2
hydra-core==1.3.2
langchain==0.2.8
langchain_community==0.2.7
langchain_core==0.2.19
langchain_google_genai==1.0.7
langchain_openai==0.1.16
langgraph==0.1.8
omegaconf==2.3.0
openai==1.35.13
pandas==2.2.2
Pillow==10.4.0
portkey_ai==1.7.0
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
tqdm==4.66.2
wandb==0.17.4
